Solved the mystery of the incompatible versions
Add some properties to build.gradle so the plugin is likely to work on more than one version of IntelliJ.

Add some graceless and repetitive logic to increase the ability to be lazy when trying to decode a cert.  Also increases the likelihood of decoding the wrong cert but then all you have to do is put the cursor in the middle of the cert and it will work.