Force virtual file refresh
Replacing the contents of a file and then invoking the plugin by right-clicking doesn't always result in the expected outcome - the plugin would work with the old contents.  Forcing a file refresh seems to solve this problem.

There may be better ways of doing this (certainly not involving synchronous file refreshes) but this should be adequate for the time being.